interation of VLC_LIBRARIES
