--- test/.gitignore
@@ -1,2 +1,0 @@
-opencvvlc
-vlc-log.txt
